refactored admin uri to default
