--- templates/admin/index.html.twig
@@ -1,17 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Hello AdminController!{% endblock %}
-
-{% block body %}
-<style>
-    .example-wrapper { margin: 1em auto; max-width: 800px; width: 95%; font: 18px/1.5 sans-serif; }
-    .example-wrapper code { background: #F5F5F5; padding: 2px 6px; }
-</style>
-
-<div class="example-wrapper">
-    <div class="news-list">
-        {% include 'blog/_list.html.twig' with {'posts': blogPosts} %}
-        {% include 'util/_pagination.html.twig' with {'currentPage': currentPage, 'pages': pages, 'path': 'admin_posts_page'} %}
-    </div>
-</div>
-{% endblock %}
