+++ templates/default/index.html.twig
@@ -1,6 +1,6 @@
 {% extends 'base.html.twig' %}
 
-{% block title %}Hello DefaultController!{% endblock %}
+{% block title %}Hello AdminController!{% endblock %}
 
 {% block body %}
 <style>
@@ -9,12 +9,9 @@
 </style>
 
 <div class="example-wrapper">
-    <h1>Hello {{ controller_name }}! ✅</h1>
-
-    This friendly message is coming from:
-    <ul>
-        <li>Your controller at <code><a href="{{ 'C:/Users/abjor/code/hum/backend/src/Controller/DefaultController.php'|file_link(0) }}">src/Controller/DefaultController.php</a></code></li>
-        <li>Your template at <code><a href="{{ 'C:/Users/abjor/code/hum/backend/templates/default/index.html.twig'|file_link(0) }}">templates/default/index.html.twig</a></code></li>
-    </ul>
+    <div class="news-list">
+        {% include 'blog/_list.html.twig' with {'posts': blogPosts} %}
+        {% include 'util/_pagination.html.twig' with {'currentPage': currentPage, 'pages': pages, 'path': 'posts_page'} %}
+    </div>
 </div>
 {% endblock %}
